rustledb: to buffer, page manager wip

+++ software-and-math-exercises/rust/rustledb/src/storage/columnar_type.rs
@@ -1,5 +1,5 @@
 pub enum ColumnarType {
-    Int(i32),
+    Int(usize),
     Varchar(String, usize),
 }
